Move password change email alert code to web service
