+++ oauth/user.go
@@ -262,35 +262,6 @@ func (s *Service) setPasswordCommon(db *bun.DB, user *model.User, password strin
 		return err
 	}
 
-	// Inform user by email password was changed
-	mg := mailgun.NewMailgun(s.cnf.Mailgun.Domain, s.cnf.Mailgun.Key)
-	sender := s.cnf.Mailgun.Sender
-	body := ""
-	email := model.NewOauthEmail(
-		user.Username,
-		"Password changed",
-		"password-changed",
-	)
-	subject := email.Subject
-	recipient := email.Recipient
-	message := mg.NewMessage(sender, subject, body, recipient)
-	message.SetTemplate(email.Template) // set mailgun template
-	err = message.AddTemplateVariable("email", recipient)
-
-	if err != nil {
-		log.ERROR.Print(err)
-	}
-
-	ctx, cancel := context.WithTimeout(context.Background(), time.Second*10)
-	defer cancel()
-
-	// Send the message with a 10 second timeout
-	_, _, err = mg.Send(ctx, message)
-
-	if err != nil {
-		log.ERROR.Print(err)
-	}
-
 	return nil
 }
